==> 01_CLEAN.ipynb <==
# Data Cleaning

### In version 2 of data cleaning, we want to introduce new columns: `home_age`, `baths`, and `garage_age`

-----
## The main objective of this notebook is to clean the data in order to have a complete dataframe (no null values). The following will be executed:
- Remove columns
- Impute missing values 
- Drop empty rows

import pandas as pd

train = pd.read_csv('./datasets/train.csv')

train.head()

train.shape

train.isnull().mean().sort_values(ascending = False).head(7)

# Removing Columns

#the columns I want to remove are in this list 
del_col = ['Pool QC',          
            'Misc Feature',     
            'Alley',            
            'Fence'  ]

train.drop(labels = del_col, axis = 1, inplace = True)

train.shape

#4 columns have been removed

train.isnull().mean().sort_values(ascending= False).head(25)

# Before I drop rows where cells are blank, I may want to impute missing values 

train.shape

# Lets look at the unique values for the columns where cell are missing

### Here are the columns where cells are missing:
    - Fireplace Qu      0.487567
    - Lot Frontage      0.160897
    - Garage Yr Blt     0.055583
    - Garage Finish     0.055583
    - Garage Qual       0.055583
    - Garage Cond       0.055583
    - Garage Type       0.055095
    - Bsmt Exposure     0.028279
    - BsmtFin Type 2    0.027304
    - Bsmt Qual         0.026816
    - BsmtFin Type 1    0.026816
    - Bsmt Cond         0.026816
    - Mas Vnr Area      0.010726
    - Mas Vnr Type      0.010726
    - Bsmt Full Bath    0.000975
    - Bsmt Half Bath    0.000975
    - BsmtFin SF 2      0.000488
    - BsmtFin SF 1      0.000488
    - Bsmt Unf SF       0.000488
    - Total Bsmt SF     0.000488
    - Garage Cars       0.000488
    - Garage Area       0.000488

# Fireplace Qu

train['Fireplace Qu'].value_counts()

### Value types for Fireplace Qu from http://jse.amstat.org/v19n3/decock/DataDocumentation.txt

#Ex	Excellent - Exceptional Masonry Fireplace
#Gd	Good - Masonry Fireplace in main level
#TA	Average - Prefabricated Fireplace in main living area or Masonry Fireplace in basement
#Fa	Fair - Prefabricated Fireplace in basement
#Po	Poor - Ben Franklin Stove
#NA	No Fireplace----- This part is missing!

# changing null cells with NA for fireplace qu
train['Fireplace Qu'].fillna('none',  inplace = True)

train['Fireplace Qu'].value_counts()

# Lot Frontage

train['Lot Frontage'].value_counts().head()

#lot frontage is numeric, how should I fill empty cells when column is numeric?

train['Lot Frontage'].fillna(train['Lot Frontage'].median(), inplace = True)

train['Lot Frontage'].isnull().sum()

# Garage Yr Blt

#numeric
train['Garage Yr Blt'].head()

train['Garage Yr Blt'].fillna(train['Garage Yr Blt'].median(), inplace = True)

train['Garage Yr Blt'].isnull().sum()

# Garage Finish

train['Garage Finish'].value_counts()


#Fin	Finished
#RFn	Rough Finished	
#Unf	Unfinished
#NA	No Garage

train['Garage Finish'].fillna('none', inplace = True)

train['Garage Finish'].isnull().sum()

# Garage Qual

train['Garage Qual'].value_counts()

#  Ex	Excellent
#        Gd	Good
#        TA	Typical/Average
#        Fa	Fair
#        Po	Poor
#        NA	No Garage

train['Garage Qual'].fillna('none', inplace = True)

train['Garage Qual'].isnull().sum()

# Garage Cond

train['Garage Cond'].value_counts()

# Ex	Excellent
#        Gd	Good
#        TA	Typical/Average
#        Fa	Fair
#        Po	Poor
#        NA	No Garage

train['Garage Cond'].fillna('none', inplace = True)

train['Garage Cond'].isnull().sum()

# Garage Type

train['Garage Type'].value_counts()

#  2Types	More than one type of garage
#        Attchd	Attached to home
#        Basment	Basement Garage
#        BuiltIn	Built-In (Garage part of house - typically has room above garage)
#        CarPort	Car Port
#        Detchd	Detached from home
#        NA	No Garage

train['Garage Type'].fillna('none', inplace = True)

train['Garage Type'].isnull().sum()

# Bsmt Exposure

train['Bsmt Exposure'].value_counts()

# Gd	Good Exposure
#        Av	Average Exposure (split levels or foyers typically score average or above)	
#        Mn	Mimimum Exposure
#        No	No Exposure
#        NA	No Basement

train['Bsmt Exposure'].fillna('none', inplace = True)

train['Bsmt Exposure'].isnull().sum()

# BsmtFin Type 2

train['BsmtFin Type 2'].value_counts()

# GLQ	Good Living Quarters
#        ALQ	Average Living Quarters
#        BLQ	Below Average Living Quarters	
#        Rec	Average Rec Room
#        LwQ	Low Quality
#        Unf	Unfinshed
#        NA	No Basement

train['BsmtFin Type 2'].fillna('none', inplace = True)

train['BsmtFin Type 2'].isnull().sum()

# Bsmt Qual

train['Bsmt Qual'].value_counts()


# Ex	Excellent (100+ inches)	
#        Gd	Good (90-99 inches)
#        TA	Typical (80-89 inches)
#        Fa	Fair (70-79 inches)
#        Po	Poor (<70 inches
#        NA	No Basement

train['Bsmt Qual'].fillna('none', inplace = True)

train['Bsmt Qual'].isnull().sum()

## At this point, this is too tedious. I realized the missing values for object columns are NA. I'm going to create a for loop that will fill missing values for me. If the column is numeric, it will fill in the data using the median. 

#my list of columns that need to be filled 
fil_col = ['Bsmt Cond',         
'Mas Vnr Area',      
'Mas Vnr Type',      
'Bsmt Full Bath',    
'Bsmt Half Bath',    
'BsmtFin SF 2',      
'BsmtFin SF 1',      
'Bsmt Unf SF',       
'Total Bsmt SF',     
'Garage Cars',       
'Garage Area',
'BsmtFin Type 1']

for col in list(fil_col):
    if train[col].dtype != 'O':
        train[col].fillna(train[col].median(), inplace = True)
    else: 
        train[col].fillna('none', inplace = True)

#check if it worked 
train.isnull().sum().sum()

train.shape

# Before we save this data, let's introduce the three new columns: `home_age`, `baths`, and `garage_age`

#creating home_age 
train['homeage'] = train['Yr Sold'] - train['Year Built']

#creating baths
train['baths'] = train['Full Bath'] + (train['Half Bath']/2)

#creating garage_age 
train['garageage'] = train['Yr Sold'] - train['Garage Yr Blt']

train.shape

train.head(1)

train.isnull().sum().sum()

# Lets save this data and call it clean_train.csv and store it in the datasets folder

train.to_csv('./datasets/clean_train.csv')
