==> 02_Dummy.ipynb <==
# This notebook will create the dummy columns for the train_clean dataset and find the dummy features with the highest correlation. 

import pandas as pd

train = pd.read_csv('./datasets/clean_train.csv')
train.head()

#just to make everything is complete
train.isnull().sum().sum()

#for loop for numeric list and object list 
numeric_col = []
object_col = []

for col in list(train):
    if train[col].dtype != 'O':
        numeric_col.append(col)
    else:
        object_col.append(col)

train = pd.get_dummies(train, columns = object_col)
train.head()

#dropping three irreleevant columns 
train.drop(labels= ['Unnamed: 0', 'Id', 'PID'], axis = 1, inplace = True)
train.head()

corr_col = []
for col in list(train):
    corr = train['SalePrice'].corr(train[col])
    if corr> 0.50:
        corr_col.append(col)

corr_col

# The Three Dummy Columns I want are: 
- `Foundation_PConc`
- `Bsmt Qual_Ex`, 
- `Kitchen Qual_Ex`

#removing three irrelevant columns from numeric_col
numeric_col.remove('Unnamed: 0')
numeric_col.remove('Id')
numeric_col.remove('PID')

#appending three dummy columns to my list 
col = numeric_col
col.append('Foundation_PConc')
col.append('Bsmt Qual_Ex')
col.append('Kitchen Qual_Ex')

#col is my new column list with all numeric column names and three dummy columns
#let's save it to a csv
train[col].head() 

train[col].shape

# Ultimately, I want to organize my train data and create a csv file with all numeric columns in addition to the dummy columns I've created. 
- Note: This dataframe is all numeric and included the three dummy columns with high correlation

### The csv file is called `df_with_dummy.csv`

#means df with dummy columns
df_wdum = train[col]
df_wdum.to_csv('./datasets/df_with_dummy.csv')
