==> 03_PreProcessing.ipynb <==
# In this notebook, I will:
- Generate polynonmial features without the dummy columns 
- Find the high correlation columns 
- Use those columns and standardize them (this is a new dataframe)
- Append the standardized dataframe with the three 3 dummy columns shown before 

from sklearn.preprocessing import StandardScaler, PolynomialFeatures
import pandas as pd

train = pd.read_csv('./datasets/df_with_dummy.csv')
train.head()

train.shape

#removing dummy column so i can create polynomial features 
numeric_col = []
for col in list(train):
    if '_' not in col: 
        numeric_col.append(col)

len(numeric_col)

#removing irrelevant columns 
numeric_col.remove('Unnamed: 0')

numeric_col.remove('SalePrice')

len(numeric_col)

# Polynomial Features

X = train[numeric_col]

poly = PolynomialFeatures(include_bias= False)
X_poly = poly.fit_transform(X)

#creating a dataframe of our polynomial features
poly_df = pd.DataFrame(X_poly, columns = poly.get_feature_names(numeric_col))

poly_df.shape

# Finding Polynomial Columns with High Correlation

train['SalePrice'].shape

#### This time, I am setting the correlation magnitude at 0.70

corr_col = []
for col in list(poly_df):
    corr = train['SalePrice'].corr(poly_df[col])
    if corr > 0.50 or corr < -.5:
        corr_col.append(col)

corr_col

len(corr_col)

poly_df[corr_col].head()

# At this point, `poly_df[corr_col]` has only columns with polynomial features. Now let's standardize it!

#I want to standardize my new polynomial features 
X = poly_df[corr_col]

ss= StandardScaler()
ss.fit(X)  #learning the means and std for each column 
poly_scaled = ss.transform(X)

poly = pd.DataFrame(poly_scaled, columns = corr_col)

poly.head()

append_col = ['Foundation_PConc','Bsmt Qual_Ex', 'Kitchen Qual_Ex', 'SalePrice']
#adding the three dummy columns and saleprice 

train[append_col].head()

mother = pd.concat(objs=[poly, train[append_col]], axis='columns')

mother.shape

# We finally made the mother dataframe! These are the features I picked. The workflow process fro preprocessing are as follows:
- Genrate polynomial features without dummy
- Find high correlation of those polynomial features
- Standardize those specific features
- Append three dummies and `SalePrice`

**This medium blog helped me streamline my preprocessing workflow.**
- https://medium.com/@samchaaa/preprocessing-why-you-should-generate-polynomial-features-first-before-standardizing-892b4326a91d

mother.to_csv('./datasets/mother.csv')

# This mother data contains:
- Non-standardized `SalePrice` column
- Three non-standardized dummy columns
- Nine polynomial and standardized columns

Besides the `SalePrice` column, these are my 9 features.
