==> 04_AMES_Gridsearch1.ipynb <==
# Ames Gridsearch
***


- Dummy varaibles are picked when corr > .50
- Polynomial Features were picked when corr > 0.50
- `train_features_for_gridsearch` is already scaled 

**Total of 253 features (250 numeric and 3 dummy columns)**

import numpy as np
import pandas as pd
import seaborn as sns

from sklearn.linear_model import ElasticNet, Lasso
from sklearn.metrics import mean_squared_error
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split, GridSearchCV

train = pd.read_csv('./datasets/train_features_for_gridsearch')
train.head()

price = pd.read_csv('./datasets/clean_train.csv')

price['SalePrice'].shape

train.shape

#lets drop Unnamed: 0 to make train have only the 23 features we want 
train.drop(labels = 'Unnamed: 0', axis = 1, inplace = True)

train.head()

train.shape

#create X and y
X= train
y = price['SalePrice']

# Train Test Split 

X_train, X_holdout, y_train, y_holdout = train_test_split(X, y, test_size = 0.3, random_state = 42)

lasso = Lasso()
lasso.fit(X_train, y_train)

# Calculate R-Squared and RMSE Scores

lasso.score(X_train, y_train)

lasso.score(X_holdout, y_holdout)

# Creating Predictions

y_train_pred = lasso.predict(X_train)
y_holdout_pred = lasso.predict(X_holdout)

def rmse(y, yhat):
    mse = mean_squared_error(y, yhat)
    return np.sqrt(mse)

#calc rmse
rmse(y_train, y_train_pred)

rmse(y_holdout, y_holdout_pred)

# Grid Search Time

check sklearn parameters guide to alter whatever you want on gridsearch

#params
lasso_params = {
    'alpha': np.linspace(0, 1, 20) , #start at .2 end at 1 and give me 8 steps between that 
    
}

#grid search
lasso_gridsearch = GridSearchCV(
    Lasso(),
    lasso_params,
    cv = 8,
    verbose = 1,
    n_jobs =2,
    return_train_score = False
)

#fit to train
lasso_gridsearch.fit(X_train, y_train)

#best params
lasso_gridsearch.best_params_

#best score
lasso_gridsearch.best_score_

#save the best model it came up with 
best_lasso = lasso_gridsearch.best_estimator_

best_lasso.coef_

#best estimator score 
best_lasso.score(X_train, y_train)

#best estimator score on holdut 
best_lasso.score(X_holdout, y_holdout)

# RMSE

#predict 
y_train_lasso = best_lasso.predict(X_train)
y_holdout_lasso = best_lasso.predict(X_holdout)

#calc rmse
print(rmse(y_train, y_train_lasso))
print(rmse(y_holdout, y_holdout_lasso))

# Conclusion 

By executing the `cross_val_score`, proceeding with the `Ridge` was the best model for us. Here are the following regression metrics for Ridge:

|Metric|Score|
|---|---|
|Train Score|.9283|
|Test Score|.9031|
|RMSE Train|21,427|
|RMSE Test|24,064|

By executing `GridSearchCV`, we used `ElasticNet` as the model since it involves both Lasso and Ridge. GridSearch will tells us the best ratio between Lasso and Ridge. After looking at the recommended parameters, the best parameters told us to have a L1 ratio of 1. We decided to perform a GridSearch for `Lasso`. We computed Lasso regression metrics. Here are the following metrics for Lasso. 

|Metric|Score|
|---|---|
|Train Score|.9280|
|Test Score|.8978|
|RMSE Train|21,467|
|RMSE Test|24,715|

**Final:** For the features we used in our regression model, it is best to proceed with the Ridge model since the overall test score and RMSE score have a better overall fit with the true sale price of homes compared to the Lasso GridSearch model.
