==> 06_TEST_SETUP.ipynb <==
# Importing clean dataframe: 3 binary columns and 39 numeric columns

import pandas as pd

d1 = pd.read_csv('./datasets/df_with_dummy.csv')

# Dummy and Numeric List

dummy_col = ['Foundation_PConc', 'Bsmt Qual_Ex', 'Kitchen Qual_Ex']

numeric_col = []
irrelevant_col = ['Unnamed: 0', 'SalePrice', 'Foundation_PConc', 'Bsmt Qual_Ex', 'Kitchen Qual_Ex']
for col in list(d1):
    if col not in irrelevant_col:
        numeric_col.append(col)
    
                                                                      
                                                                      

len(numeric_col)

# Polynomial for Train

from sklearn.preprocessing import StandardScaler, PolynomialFeatures

X = d1[numeric_col]
poly = PolynomialFeatures(include_bias= False)
X_train_poly = poly.fit_transform(X)

#creating a dataframe of our polynomial features
poly_df = pd.DataFrame(X_train_poly, columns = poly.get_feature_names(numeric_col))

## `poly_df` is the mother of polynomial features 

poly_df.head()

corr_col = ['Overall Qual',
 'Year Built',
 'Year Remod/Add',
 'Mas Vnr Area',
 'Total Bsmt SF',
 '1st Flr SF',
 'Gr Liv Area',
 'Full Bath',
 'TotRms AbvGrd',
 'Garage Yr Blt',
 'Garage Cars',
 'Garage Area',
 'homeage',
 'baths',
 'garageage',
 'Lot Frontage Overall Qual',
 'Lot Frontage Mas Vnr Area',
 'Lot Frontage Gr Liv Area',
 'Lot Frontage Full Bath',
 'Lot Frontage Garage Cars',
 'Lot Frontage Garage Area',
 'Lot Frontage baths',
 'Lot Area Overall Qual',
 'Lot Area Garage Cars',
 'Overall Qual^2',
 'Overall Qual Overall Cond',
 'Overall Qual Year Built',
 'Overall Qual Year Remod/Add',
 'Overall Qual Mas Vnr Area',
 'Overall Qual BsmtFin SF 1',
 'Overall Qual Total Bsmt SF',
 'Overall Qual 1st Flr SF',
 'Overall Qual Gr Liv Area',
 'Overall Qual Full Bath',
 'Overall Qual Bedroom AbvGr',
 'Overall Qual Kitchen AbvGr',
 'Overall Qual TotRms AbvGrd',
 'Overall Qual Fireplaces',
 'Overall Qual Garage Yr Blt',
 'Overall Qual Garage Cars',
 'Overall Qual Garage Area',
 'Overall Qual Yr Sold',
 'Overall Qual baths',
 'Overall Cond Total Bsmt SF',
 'Overall Cond 1st Flr SF',
 'Overall Cond Gr Liv Area',
 'Overall Cond Garage Cars',
 'Overall Cond Garage Area',
 'Overall Cond baths',
 'Year Built^2',
 'Year Built Year Remod/Add',
 'Year Built Mas Vnr Area',
 'Year Built Total Bsmt SF',
 'Year Built 1st Flr SF',
 'Year Built Gr Liv Area',
 'Year Built Full Bath',
 'Year Built TotRms AbvGrd',
 'Year Built Garage Yr Blt',
 'Year Built Garage Cars',
 'Year Built Garage Area',
 'Year Built Yr Sold',
 'Year Built homeage',
 'Year Built baths',
 'Year Built garageage',
 'Year Remod/Add^2',
 'Year Remod/Add Mas Vnr Area',
 'Year Remod/Add Total Bsmt SF',
 'Year Remod/Add 1st Flr SF',
 'Year Remod/Add Gr Liv Area',
 'Year Remod/Add Full Bath',
 'Year Remod/Add TotRms AbvGrd',
 'Year Remod/Add Garage Yr Blt',
 'Year Remod/Add Garage Cars',
 'Year Remod/Add Garage Area',
 'Year Remod/Add Yr Sold',
 'Year Remod/Add homeage',
 'Year Remod/Add baths',
 'Year Remod/Add garageage',
 'Mas Vnr Area Total Bsmt SF',
 'Mas Vnr Area 1st Flr SF',
 'Mas Vnr Area Gr Liv Area',
 'Mas Vnr Area Full Bath',
 'Mas Vnr Area Bedroom AbvGr',
 'Mas Vnr Area TotRms AbvGrd',
 'Mas Vnr Area Fireplaces',
 'Mas Vnr Area Garage Yr Blt',
 'Mas Vnr Area Garage Cars',
 'Mas Vnr Area Garage Area',
 'Mas Vnr Area Yr Sold',
 'Mas Vnr Area baths',
 'BsmtFin SF 1 Full Bath',
 'BsmtFin SF 1 Garage Cars',
 'BsmtFin SF 1 baths',
 'Total Bsmt SF 1st Flr SF',
 'Total Bsmt SF Gr Liv Area',
 'Total Bsmt SF Full Bath',
 'Total Bsmt SF Bedroom AbvGr',
 'Total Bsmt SF TotRms AbvGrd',
 'Total Bsmt SF Fireplaces',
 'Total Bsmt SF Garage Yr Blt',
 'Total Bsmt SF Garage Cars',
 'Total Bsmt SF Garage Area',
 'Total Bsmt SF Yr Sold',
 'Total Bsmt SF baths',
 '1st Flr SF^2',
 '1st Flr SF Gr Liv Area',
 '1st Flr SF Full Bath',
 '1st Flr SF TotRms AbvGrd',
 '1st Flr SF Fireplaces',
 '1st Flr SF Garage Yr Blt',
 '1st Flr SF Garage Cars',
 '1st Flr SF Garage Area',
 '1st Flr SF Yr Sold',
 '1st Flr SF baths',
 'Gr Liv Area^2',
 'Gr Liv Area Full Bath',
 'Gr Liv Area TotRms AbvGrd',
 'Gr Liv Area Fireplaces',
 'Gr Liv Area Garage Yr Blt',
 'Gr Liv Area Garage Cars',
 'Gr Liv Area Garage Area',
 'Gr Liv Area Yr Sold',
 'Gr Liv Area baths',
 'Full Bath^2',
 'Full Bath TotRms AbvGrd',
 'Full Bath Fireplaces',
 'Full Bath Garage Yr Blt',
 'Full Bath Garage Cars',
 'Full Bath Garage Area',
 'Full Bath Yr Sold',
 'Full Bath baths',
 'Bedroom AbvGr Garage Cars',
 'Bedroom AbvGr Garage Area',
 'Kitchen AbvGr Garage Cars',
 'Kitchen AbvGr Garage Area',
 'Kitchen AbvGr homeage',
 'Kitchen AbvGr garageage',
 'TotRms AbvGrd Fireplaces',
 'TotRms AbvGrd Garage Yr Blt',
 'TotRms AbvGrd Garage Cars',
 'TotRms AbvGrd Garage Area',
 'TotRms AbvGrd Yr Sold',
 'TotRms AbvGrd baths',
 'Fireplaces Garage Cars',
 'Fireplaces Garage Area',
 'Fireplaces baths',
 'Garage Yr Blt^2',
 'Garage Yr Blt Garage Cars',
 'Garage Yr Blt Garage Area',
 'Garage Yr Blt Yr Sold',
 'Garage Yr Blt homeage',
 'Garage Yr Blt baths',
 'Garage Yr Blt garageage',
 'Garage Cars^2',
 'Garage Cars Garage Area',
 'Garage Cars Yr Sold',
 'Garage Cars baths',
 'Garage Area^2',
 'Garage Area Yr Sold',
 'Garage Area baths',
 'Yr Sold homeage',
 'Yr Sold baths',
 'Yr Sold garageage',
 'baths^2']
#list is from another notebook- but these are the columns that have a correlation higher than .50

len(poly_df[corr_col].columns)

# Train Scale Time 

#I want to standardize my new polynomial features 
X = poly_df[corr_col]

ss= StandardScaler()
ss.fit(X)  #learning the means and std for each column 
poly_scaled = ss.transform(X)

train_poly = pd.DataFrame(poly_scaled, columns = corr_col)

train_poly.shape

### When we transform test, we must use  poly_test_scaled = ss.transform(X) then test_poly = pd.DataFrame(poly_test_scaled, columns = corr_col)

d1[dummy_col].shape

train_sc_dum = train_poly.join(d1[dummy_col])

train_sc_dum.shape

train_sc_dum.head()

train_sc_dum.to_csv('./datasets/train_features_for_gridsearch')

# Model 

from sklearn import metrics
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.linear_model import LinearRegression, LassoCV, RidgeCV, ElasticNetCV
from sklearn.preprocessing import StandardScaler, PowerTransformer
from sklearn.metrics import r2_score

import numpy as np

#creating our x and y
X = train_sc_dum
y = d1['SalePrice']

#train test split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

#instantiate the models
lr = LinearRegression()
lasso = LassoCV()
ridge = RidgeCV()
elastic = ElasticNetCV()

# Cross Val Scores

cross_val_score(lr, X_train, y_train, cv= 10).mean()

cross_val_score(ridge, X_train, y_train, cv= 10).mean()

cross_val_score(elastic, X_train, y_train, cv= 10).mean()

cross_val_score(lasso, X_train, y_train, cv= 10).mean()

# Proceed with Ridge

ridge.fit(X_train, y_train)

#cv scores
cross_val_score(ridge, X_train, y_train, cv=10).mean()

#train score
ridge.score(X_train, y_train)

#test score
ridge.score(X_test, y_test)

from sklearn.metrics import mean_squared_error
import math

#train prediction 
train_pred = ridge.predict(X_train)

#mse train
mean_squared_error(y_train, train_pred)

#import numpy as np

#rmse train
np.sqrt(mean_squared_error(y_train, train_pred))

#test prediction 
test_pred = ridge.predict(X_test)

#mse test
mean_squared_error(y_test, test_pred)

#rmse test
np.sqrt(mean_squared_error(y_test, test_pred))

# Import Clean Test.csv data

test = pd.read_csv('./datasets/clean_test.csv')

test.head()

test[numeric_col].shape

test[dummy_col].shape

# Polynomial Transformation of Test

X = test[numeric_col]
poly = PolynomialFeatures(include_bias= False)
X_test_poly = poly.fit_transform(X)

#creating a dataframe of our polynomial features
poly_df = pd.DataFrame(X_test_poly, columns = poly.get_feature_names(numeric_col))

poly_df.shape

#using same list from model 
poly_df[corr_col].shape

# Test Scale Time

#I want to standardize my new polynomial features 
X = poly_df[corr_col]

poly_test_scaled = ss.transform(X)

test_poly = pd.DataFrame(poly_test_scaled, columns = corr_col)

test_poly.shape

# Lets append the dummies

test[dummy_col].shape

test_features = test_poly.join(test[dummy_col])

test_features.shape

# Predict using Ridge

pred = ridge.predict(test_features)

# Before we transfer our predictions, lets make sure our model is right
- In the `Ames_GridSearch1` notebook, we will grid search using elastic net to see best fit and determine if we should consider the Lasso and or Ridge model. 

*Note: Please look at the `Ames_GridSearch1` notebook to see how we executed our GrideSearch model*

# Model Conclusion

By executing the `cross_val_score`, proceeding with the `Ridge` was the best model for us. Here are the following regression metrics for Ridge:

|Metric|Score|
|---|---|
|Train Score|.9283|
|Test Score|.9031|
|RMSE Train|21,427|
|RMSE Test|24,064|

By executing `GridSearchCV`, we used `ElasticNet` as the model since it involves both Lasso and Ridge. GridSearch will tells us the best ratio between Lasso and Ridge. After looking at the recommended parameters, we computed regression metrics. Here are the following metrics for elastic. 

|Metric|Score|
|---|---|
|Train Score|.9280|
|Test Score|.8978|
|RMSE Train|21,467|
|RMSE Test|24,711|

**Final:** For the features we used in our regression model, it is best to proceed with the ridge model since the overall test score and RMSE score have a better overall fit with the true sale price of homes. 

# Create CSV to import results

test['Id'].shape

pred.shape

submission = pd.DataFrame({
    'Id': test['Id'],
    'SalePrice': pred
})

submission.to_csv('./datasets/submissionv_thattooklong4.csv', index= False)
